Refactor and fix initial field positioning logic
This commit addresses two issues with the initial generation of the preview from background images:

1.  The editable fields are now positioned using the "auto-organize" logic on initial generation, ensuring a consistent layout whether the user has just loaded the data or has manually triggered the auto-organize feature.
2.  The positioning logic for the third field (the side label) has been corrected. It is now properly sized and positioned against the right border of the image after its 270-degree rotation.

To achieve this, the `autoArrangeFields` logic was refactored out of the `FieldPositioner` component and into a new utility file, `src/utils/autoArrange.js`. This allows the logic to be reused in `HomePage.jsx` for the initial setup of the fields.
